+++ archinstall/lib/disk/btrfs.py
